remove text about sharing no characteristics
Unfortunately, sharing no characteristics is nearly impossible on
the web today including it would frustrate users to grant explicit
consent for every possible browser fingerprinting characteristic
necessary to achieve compatibility across the web. I've updated
the text here so that it conveys the original intent of minimizing
uniqueness while also removing text that can't really be adhered
with for a variety of complex reasons.