Avoid saying that safety systems have to describe how to exploit them.
The text I originally wrote here got interpreted as saying that services
have to explain all the details of how they use data to fight fraud,
which of course would help the fraudsters avoid detection. Instead, we
meant just that the service should say what kinds of things it's
collecting in order to protect people.

I also got a comment on how abrupt this paragraph is, since it just
restricts the safety-related data collection without saying that this
purpose is reasonable in the first place.