No excuses for fingerprinting (#275)
The notion that fingerprinting might not be economically feasible
doesn't really help anyone.  On the one hand, it just tends to create
an excuse for lazy web designers who might choose to neglect
fingerprinting protection on the basis that "it won't be viable to
exploit this" (tip: it probably is viable).  On the other, the
admonition about people who didn't ask to be treated that way is already
implied.y

Also, the document has enough words in it already.  No point playing
both sides here.

I considered saying something more in place of this, but the paragraph
is already complete as is.  That is, it talks about the challenges
inherent with fingerprinting in that it disproportionately affects
those smaller groupings that tend to form around less mainstream
choices or characteristics.